requirements.txt
networkx==2.4
requests==2.22.0
colorthief==0.2.1
pandas==1.1.4
streamlit==0.80.0
Pillow==8.2.0
